Drop Fetcher class; pool-as-arg + DownloadPolicy on download_all
Per PR #101 review discussion: ``Fetcher`` was a parameter object pretending
to be a class. The actual concerns split cleanly into:

  * concurrency = a thread pool, not a "fetcher"
  * policy = three booleans, frozen dataclass (``DownloadPolicy``)
  * orchestration = a free function (``_drive``)

So ``Fetcher`` deletes. ``Source.download_all`` becomes the only public fetch
entry point, with optional ``pool`` and ``policy`` kwargs:

  src.download_all(dest)                              # private pool, default policy
  src.download_all(dest, pool=pool, policy=policy)    # caller-owned pool, shared policy

The CLI now builds one ``ThreadPoolExecutor`` via ExitStack and hands it to
every Source in the loop — the actual shared-pool benefit we'd been talking
about but never realized in the previous Fetcher-class design (which built
a fresh pool inside ``Fetcher.fetch_all`` on every call).

## Cycle gone, ``_types.py`` deleted

The previous redesign needed ``_types.py`` to break the source.py ↔ fetcher.py
import cycle (Source's ``__init__`` had a ``Fetcher()`` default). With Fetcher
gone, no cycle remains; every type lives in source.py:

  * ``RemoteFile`` (manifest row, internal)
  * ``ChecksumError``, ``sha256_of`` (transport helpers)
  * ``DownloadPolicy`` (frozen config)
  * ``FetchResult``, ``FetchReport`` (outcome types)
  * ``_resolve_dest``, ``_already_complete``, ``_maybe_unarchive``,
    ``_fetch_one``, ``_drive`` (free functions, module-private)
  * ``Source`` (the ABC)

One file, ~580 lines, no internal cycle.

## SIGINT-cancel-futures preserved

The painstakingly-debugged ``shutdown(wait=False, cancel_futures=True)`` from
PR #88 is still here — but only on the path where ``download_all`` *owns* the
pool (no ``pool=`` kwarg passed). When a caller passes their own pool, the
pool's lifetime is the caller's responsibility (typically an ExitStack in the
CLI, which still does ``shutdown(wait=True)`` on normal exit; SIGINT
propagates out before the stack unwinds). The SIGINT child-test uses the
private-pool path so the regression guard is unchanged.

## Backends

Constructors lose their ``fetcher=`` kwarg. ``HTTPSource(urls, *, client=...)``,
``PhysioNetSource(base_url, ...)``, ``FsspecSource(root)`` are simpler. Each
backend's ``_list_files`` and ``_fetch`` bodies are unchanged.

## CLI

```python
policy = DownloadPolicy(continue_on_error=..., do_overwrite=...)
with ExitStack() as stack:
    pool = stack.enter_context(ThreadPoolExecutor(max_workers=cfg.concurrency))
    for source in sources:
        stack.enter_context(source)
    for source in sources:
        report = source.download_all(raw_input_dir, pool=pool, policy=policy)
        ...
```

Net delta vs the previous ``Fetcher``-based version of #101: -111 lines, one
fewer module, no import cycle. Cleaner public surface (``DownloadPolicy``
replaces ``Fetcher`` in the __all__).

All 126 non-integration tests pass. Pre-commit clean.
