+++ src/MEDS_extract/download/dispatch.py
@@ -17,11 +17,10 @@ from typing import TYPE_CHECKING
 from .backends import FsspecSource, HTTPSource, PhysioNetSource
 
 if TYPE_CHECKING:
-    from .fetcher import Fetcher
     from .source import Source
 
 
-def source_from_config(cfg: dict, *, fetcher: Fetcher | None = None) -> Source:
+def source_from_config(cfg: dict) -> Source:
     """Construct one :class:`Source` from a single ``sources:`` list entry.
 
     Each entry is a dict carrying at minimum a ``type:`` field. Remaining keys are
@@ -70,8 +69,6 @@ def source_from_config(cfg: dict, *, fetcher: Fetcher | None = None) -> Source:
     source_type = cfg.pop("type", None)
     if source_type is None:
         raise ValueError(f"Source config is missing a 'type:' key. Got: {cfg}")
-    if fetcher is not None:
-        cfg["fetcher"] = fetcher
 
     match source_type:
         case "physionet":
@@ -86,7 +83,7 @@ def source_from_config(cfg: dict, *, fetcher: Fetcher | None = None) -> Source:
             )
 
 
-def sources_from_spec(spec: dict, key: str = "dataset", *, fetcher: Fetcher | None = None) -> list[Source]:
+def sources_from_spec(spec: dict, key: str = "dataset") -> list[Source]:
     """Read a full MESSY ``sources:`` block and return the configured + common sources.
 
     The ``key`` argument selects which bucket of sources to pull — ``"dataset"`` (the
@@ -128,4 +125,4 @@ def sources_from_spec(spec: dict, key: str = "dataset", *, fetcher: Fetcher | No
     sources_block = spec.get("sources", {}) or {}
     configured = list(sources_block.get(key, []) or [])
     common = list(sources_block.get("common", []) or [])
-    return [source_from_config(c, fetcher=fetcher) for c in configured + common]
+    return [source_from_config(c) for c in configured + common]
